Refactor project and update ide
Moved java files to set to packages to better organize the project
