+++ .idea/vcs.xml
@@ -4,4 +4,3 @@
     <mapping directory="$PROJECT_DIR$" vcs="Git" />
   </component>
 </project>
-
