+++ app/src/main/java/ivansued/com/selfdestructingmessageapp/utils/ImageResizer.java
@@ -1,4 +1,4 @@
-package ivansued.com.selfdestructingmessageapp;
+package ivansued.com.selfdestructingmessageapp.utils;
 
 import java.io.ByteArrayOutputStream;
