+++ app/src/main/res/layout/activity_edit_friends.xml
@@ -2,7 +2,7 @@
                 xmlns:tools="http://schemas.android.com/tools"
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
-                tools:context="ivansued.com.selfdestructingmessageapp.EditFriendsActivity">
+                tools:context="ivansued.com.selfdestructingmessageapp.ui.EditFriendsActivity">
 
     <ListView
         android:layout_width="match_parent"
