+++ app/src/main/res/layout/activity_recipients.xml
@@ -6,7 +6,7 @@
                 android:paddingRight="@dimen/activity_horizontal_margin"
                 android:paddingTop="@dimen/activity_vertical_margin"
                 android:paddingBottom="@dimen/activity_vertical_margin"
-                tools:context="ivansued.com.selfdestructingmessageapp.RecipientsActivity">
+                tools:context="ivansued.com.selfdestructingmessageapp.ui.RecipientsActivity">
 
     <ListView
         android:layout_width="match_parent"
