+++ app/src/main/res/layout/activity_view_image.xml
@@ -2,7 +2,7 @@
                 xmlns:tools="http://schemas.android.com/tools"
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
-                tools:context="ivansued.com.selfdestructingmessageapp.ViewImageActivity">
+                tools:context="ivansued.com.selfdestructingmessageapp.ui.ViewImageActivity">
 
     <ImageView
         android:layout_width="match_parent"
